+++ .env.example
@@ -23,7 +23,7 @@ SUPERMARKT_RESULT_RETENTION_HOURS=168
 SUPERMARKT_TIMEOUT_SECONDS=25
 SUPERMARKT_MARKTGURU_PAGE_SIZE=500
 SUPERMARKT_MAX_WORKERS=8
-# Leer verwendet automatisch supermarkt-preisvergleich/<Version>.
+# Leer verwendet automatisch korb-klar/<Version>.
 SUPERMARKT_USER_AGENT=
 
 # Filialzuordnungen. Die Angebotsdaten selbst bleiben frisch.
