+++ pyproject.toml
@@ -3,9 +3,9 @@ requires = ["setuptools>=68"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "supermarkt-preisvergleich"
-version = "0.0.1"
-description = "Self-hosted Docker service for comparing regional supermarket offers in Germany."
+name = "korb-klar"
+version = "0.1.0"
+description = "KorbKlar: self-hosted comparison of regional supermarket offers in Germany."
 readme = "README.md"
 requires-python = ">=3.12"
 license = "BSD-3-Clause"
@@ -27,6 +27,9 @@ package-dir = {"" = "src"}
 
 [tool.setuptools.packages.find]
 where = ["src"]
+
+[tool.setuptools.package-data]
+supermarkt = ["static/*.html", "static/*.css", "static/*.js"]
 
 [tool.pytest.ini_options]
 addopts = "-q"
