+++ src/supermarkt/api_routes.py
@@ -1,0 +1,53 @@
+from __future__ import annotations
+
+from typing import Any, Literal
+
+from fastapi import APIRouter, Depends, HTTPException, Query, Request
+
+from .access import build_result_url, proxy_page_images, require_api_auth, verify_result_token
+from .api_models import SupermarketRequest
+from .loyalty import normalize_program_ids
+from .models import ToolError
+from . import runtime
+
+router = APIRouter()
+
+
+@router.post(
+    "/api/v1/compare",
+    operation_id="supermarkt_preisvergleich",
+    summary="Supermarktangebote vergleichen",
+    description="Lädt aktuelle regionale Supermarktangebote und gibt einen result_url zur interaktiven Liste zurück.",
+)
+def supermarket_compare(request_data: SupermarketRequest, request: Request, _: None = Depends(require_api_auth)) -> dict[str, Any]:
+    try:
+        engine = runtime.get_engine()
+        snapshot, from_cache = engine.snapshot(request_data.postal_code, request_data.aldi_region, request_data.refresh)
+        page = engine.page(snapshot, filter_text=request_data.filter_text, retailer=request_data.retailer, page=request_data.page, page_size=request_data.page_size, view=request_data.view, loyalty_programs=tuple(request_data.loyalty_programs), sort=request_data.sort, include_image_urls=False)
+        page["status"] = "ok"
+        page["from_cache"] = from_cache
+        page["result_url"] = build_result_url(request, snapshot["search_id"], tuple(request_data.loyalty_programs))
+        return page
+    except ToolError as exc:
+        raise HTTPException(status_code=502, detail=str(exc)) from exc
+
+
+@router.get("/api/results/{search_id}", include_in_schema=False)
+def result_data(
+    search_id: str,
+    token: str = Query(default=""),
+    q: str = Query(default="", max_length=120),
+    retailer: str = Query(default="", max_length=60),
+    page: int = Query(default=1, ge=1, le=10000),
+    page_size: int = Query(default=100, ge=1, le=100),
+    view: Literal["best_only", "all"] = Query(default="best_only"),
+    loyalty: str = Query(default="", max_length=500),
+    sort: Literal["price", "unit_price", "retailer", "product"] = Query(default="price"),
+) -> dict[str, Any]:
+    verify_result_token(search_id, token)
+    try:
+        engine = runtime.get_engine()
+        snapshot = engine.by_id(search_id)
+        return proxy_page_images(engine.page(snapshot, filter_text=q, retailer=retailer, page=page, page_size=page_size, view=view, loyalty_programs=normalize_program_ids(loyalty.split(",")), sort=sort, include_image_urls=True))
+    except ToolError as exc:
+        raise HTTPException(status_code=410, detail=str(exc)) from exc
