+++ src/supermarkt/assets.py
@@ -1,9 +1,9 @@
 from __future__ import annotations
 
 
-FAVICON_SVG = b'''<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 64 64">
-<rect width="64" height="64" rx="14" fill="#0b57d0"/>
-<path d="M18 25h28l-3 22H21l-3-22Z" fill="none" stroke="#fff" stroke-width="4" stroke-linejoin="round"/>
-<path d="M24 25c0-6 3-10 8-10s8 4 8 10" fill="none" stroke="#fff" stroke-width="4" stroke-linecap="round"/>
-<path d="M27 32v8M37 32v8" stroke="#fff" stroke-width="4" stroke-linecap="round"/>
+FAVICON_SVG = b'''<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 64 64" role="img" aria-label="KorbKlar">
+<rect width="64" height="64" rx="15" fill="#116149"/>
+<path d="M17 27h30l-4 22H21l-4-22Z" fill="#fff" opacity=".96"/>
+<path d="M23 27c1-8 4-12 9-12s8 4 9 12" fill="none" stroke="#fff" stroke-width="5" stroke-linecap="round"/>
+<path d="m26 37 4 4 9-10" fill="none" stroke="#116149" stroke-width="4" stroke-linecap="round" stroke-linejoin="round"/>
 </svg>'''
