+++ src/supermarkt/static/home.css
@@ -1,0 +1,8 @@
+:root{--bg:#f4f7f5;--panel:#fff;--text:#14201b;--muted:#63716a;--line:#dce5e0;--accent:#116149;--error:#b42318;--shadow:0 12px 38px rgba(17,97,73,.10)}
+@media(prefers-color-scheme:dark){:root{--bg:#111318;--panel:#191c22;--text:#f1f3f5;--muted:#aab1bc;--line:#303641;--accent:#72d5a6;--error:#ff938a;--shadow:none}}
+*{box-sizing:border-box} body{margin:0;min-height:100vh;display:grid;place-items:center;background:var(--bg);color:var(--text);font:16px/1.5 system-ui,-apple-system,BlinkMacSystemFont,"Segoe UI",sans-serif;padding:24px}
+.card{width:min(620px,100%);background:var(--panel);border:1px solid var(--line);border-radius:18px;padding:28px;box-shadow:var(--shadow)}
+h1{font-size:30px;line-height:1.15;margin:0 0 10px} p{margin:0 0 24px;color:var(--muted)}
+form{display:grid;grid-template-columns:1fr auto;gap:10px} input,button{font:inherit;border-radius:12px;padding:13px 14px} input{min-width:0;border:1px solid var(--line);background:var(--panel);color:var(--text);font-size:18px;letter-spacing:.05em} button{border:1px solid var(--accent);background:var(--accent);color:#fff;font-weight:700;cursor:pointer} button:disabled{opacity:.65;cursor:wait}
+.hint{font-size:13px;color:var(--muted);margin-top:14px} .error{margin:0 0 16px;padding:10px 12px;border:1px solid color-mix(in srgb,var(--error) 45%,var(--line));border-radius:10px;color:var(--error)}
+@media(max-width:520px){.card{padding:22px} form{grid-template-columns:1fr}}
