+++ src/supermarkt/static/home.js
@@ -1,0 +1,1 @@
+document.getElementById("searchForm").addEventListener("submit",()=>{const b=document.getElementById("submitButton");b.disabled=true;b.textContent="Angebote werden geladen …";});
