+++ src/supermarkt/static/results.html
@@ -1,0 +1,22 @@
+<!doctype html>
+<html lang="de">
+<head>
+<meta charset="utf-8">
+<meta name="viewport" content="width=device-width,initial-scale=1">
+<meta name="color-scheme" content="light dark">
+<link rel="icon" type="image/svg+xml" href="/favicon.svg">
+<link rel="alternate icon" href="/favicon.ico">
+<meta name="description" content="KorbKlar vergleicht aktuelle regionale Supermarktangebote.">
+<meta name="theme-color" content="#116149">
+<title>KorbKlar – Angebote</title>
+<link rel="stylesheet" href="/static/results.css">
+</head>
+<body data-search-id="__SEARCH_ID__" data-token="__RESULT_TOKEN__" data-loyalty="__LOYALTY_PROGRAMS__"><div class="wrap">
+<div class="top"><div class="titlebar"><h1>Korb<span style="color:var(--accent)">Klar</span></h1><a class="newSearch" href="/">Neue Suche</a></div><div class="meta" id="meta">Lade gespeicherten Vergleich …</div>
+<div class="controls"><input id="q" type="search" placeholder="Produkt oder Marke filtern"><select id="sort"><option value="price">Preis mit Auswahl</option><option value="unit_price">Grundpreis mit Auswahl</option><option value="retailer">Händler</option><option value="product">Produktname</option></select></div><div class="viewTabs" role="tablist" aria-label="Dubletten"><button type="button" class="viewTab active" id="viewBest" data-view="best_only" role="tab" aria-selected="true">Günstigste</button><button type="button" class="viewTab" id="viewAll" data-view="all" role="tab" aria-selected="false">Teurere Dubletten einblenden</button></div>
+<div class="loyaltyBox" id="loyaltyBox" hidden><div class="loyaltyHead"><div class="loyaltyTitle">Bonusprogramme</div><div class="loyaltyActions"><button type="button" id="loyaltyAll">Alle</button><button type="button" id="loyaltyNone">Keine</button></div></div><div class="loyaltyPrograms" id="loyaltyPrograms"></div><div class="loyaltyNote" id="loyaltyNote"></div></div>
+<div class="chips" id="chips"></div><div class="stats" id="stats"></div></div>
+<details id="warningsBox" hidden><summary>Hinweise oder Fehler</summary><div id="warnings"></div></details>
+<div class="table" id="table"><div class="row head"><div>Bild</div><div class="retailer">Händler</div><div>Produkt</div><div>Ohne Bonus</div><div>Mit Auswahl</div><div>Packung / Grundpreis</div><div>Gültig</div></div><div id="rows"></div></div>
+<div id="sentinel">Lade Angebote …</div></div>
+<script src="/static/results.js" defer></script></body></html>
