+++ src/supermarkt/static/results.js
@@ -1,0 +1,13 @@
+const pageData=document.body.dataset;
+const searchId=pageData.searchId;
+const token=pageData.token;
+let page=1, loading=false, done=false, retailer="", view="best_only", debounce=null, programsInitialized=false, requestSeq=0, controller=null;
+const selectedPrograms=new Set((pageData.loyalty||"").split(",").filter(Boolean));
+const $=id=>document.getElementById(id); const esc=s=>String(s??"").replace(/[&<>"']/g,c=>({"&":"&amp;","<":"&lt;",">":"&gt;","\"":"&quot;","'":"&#39;"}[c]));
+function syncLoyaltyUrl(){const u=new URL(location.href); const value=[...selectedPrograms].sort().join(","); value?u.searchParams.set("loyalty",value):u.searchParams.delete("loyalty"); history.replaceState(null,"",u);}
+function query(reset=false){if(reset){requestSeq++;controller?.abort();loading=false;page=1;done=false;$("rows").innerHTML="";} if(loading||done)return; loading=true; const seq=requestSeq; controller=new AbortController(); $("sentinel").textContent="Lade Angebote …"; const p=new URLSearchParams({token,page:String(page),page_size:"100",q:$("q").value,retailer,sort:$("sort").value,view,loyalty:[...selectedPrograms].sort().join(",")}); fetch(`/api/results/${encodeURIComponent(searchId)}?${p}`,{cache:"no-store",signal:controller.signal}).then(async r=>{if(!r.ok)throw new Error((await r.json().catch(()=>({}))).detail||`HTTP ${r.status}`); return r.json();}).then(data=>{if(seq!==requestSeq)return;render(data,reset);page=data.page+1;done=!data.has_next;$("sentinel").textContent=done?"Alle passenden Angebote geladen.":"Weitere Angebote werden beim Scrollen geladen …";}).catch(e=>{if(e.name==="AbortError")return;$("sentinel").innerHTML=`<span class="error">${esc(e.message)}</span>`;done=true;}).finally(()=>{if(seq===requestSeq)loading=false;});}
+function render(data,reset){$(`table`).classList.toggle("single-retailer",Boolean(retailer)); $(`meta`).textContent=`PLZ ${data.postal_code} · ${data.source_offer_count} Quellen-Treffer · Cache ${Math.floor(data.cache_age_seconds/60)} Min. alt`; const hidden=Number(data.hidden_count||0); const duplicateText=view==="all"?`${hidden} teurere Dubletten eingeblendet`:`${hidden} teurere Dubletten ausgeblendet`; $(`stats`).innerHTML=`<span>${data.filtered_offer_count} passende Treffer</span><span>${duplicateText}</span>`; $("viewBest").classList.toggle("active",view==="best_only");$("viewBest").setAttribute("aria-selected",String(view==="best_only"));$("viewAll").classList.toggle("active",view==="all");$("viewAll").setAttribute("aria-selected",String(view==="all"));$("viewAll").textContent=hidden?`Teurere Dubletten einblenden · ${hidden}`:"Teurere Dubletten einblenden"; renderChips(data.retailer_counts); renderPrograms(data.available_loyalty_programs||[],data.loyalty_note||""); const warnings=data.warnings||[]; $(`warningsBox`).hidden=!warnings.length; $(`warnings`).innerHTML=warnings.map(x=>`<div>${esc(x)}</div>`).join(""); const frag=document.createDocumentFragment(); for(const o of data.offers){const row=document.createElement("div");row.className="row"; const img=o.image_url?`<img class="thumb" loading="lazy" src="${esc(o.image_url)}" onerror="this.style.visibility='hidden'">`:`<div class="thumb"></div>`; const source=o.source_url?`<a href="${esc(o.source_url)}" target="_blank" rel="noreferrer">${esc(o.product)}</a>`:esc(o.product); const saved=(o.loyalty_savings||0)>0.004; const selectedInfo=[o.loyalty_benefit,saved?`spart ${o.loyalty_savings_text}`:""].filter(Boolean).map(esc).join(" · "); const selectedUnit=o.selected_unit_price?`<div class="small">entspricht ${esc(o.selected_unit_price)}</div>`:""; const checkoutDiff=saved&&Math.abs((o.checkout_price??o.regular_price)-(o.effective_price??o.regular_price))>0.004?`<div class="small">${esc(o.checkout_price_text)} an der Kasse</div>`:""; const regularNote=o.regular_comparison?`<div class="small priceNote">${esc(o.regular_comparison)}</div>`:""; const selectedNote=o.selected_comparison?`<div class="small priceNote">${esc(o.selected_comparison)}</div>`:""; row.innerHTML=`<div>${img}</div><div class="retailer"><strong>${esc(o.retailer)}</strong></div><div class="productblock"><div class="product">${source}</div><div class="small">${esc(o.description)}</div></div><div class="regularPrice price ${esc(o.regular_comparison_state)}">${esc(o.regular_price_text)}${regularNote}</div><div class="selectedPrice price ${esc(o.selected_comparison_state)}${saved?" saving":""}">${esc(o.effective_price_text)}${checkoutDiff}${selectedUnit}${selectedInfo?`<div class="${saved?"savingText":"small"}">${selectedInfo}</div>`:""}${selectedNote}</div><div class="details"><div>${esc(o.pack)}</div><div class="small">${esc(o.unit_price)}</div></div><div class="validity small">${esc(o.validity)}</div>`; frag.appendChild(row);} $("rows").appendChild(frag);}
+function renderChips(counts){const box=$("chips"); const entries=Object.entries(counts||{}); box.innerHTML=""; const add=(name,count)=>{const b=document.createElement("button");b.className="chip"+(retailer===name?" active":"");b.textContent=name?`${name} · ${count}`:`Alle Händler · ${entries.reduce((a,[,n])=>a+n,0)}`;b.onclick=()=>{retailer=name;query(true)};box.appendChild(b)}; add("",0); entries.sort((a,b)=>a[0].localeCompare(b[0],"de")).forEach(([n,c])=>add(n,c));}
+function renderPrograms(programs,note){const box=$("loyaltyBox"), list=$("loyaltyPrograms"); box.hidden=!programs.length; $("loyaltyNote").textContent=note; if(programsInitialized)return; programsInitialized=true; list.innerHTML=""; for(const program of programs){const label=document.createElement("label"); const input=document.createElement("input"); input.type="checkbox"; input.dataset.program=program.id; input.checked=selectedPrograms.has(program.id); input.onchange=()=>{input.checked?selectedPrograms.add(program.id):selectedPrograms.delete(program.id);syncLoyaltyUrl();query(true)}; const text=document.createElement("span"); text.textContent=program.label; const count=Number(program.priced_offer_count||0); label.title=(program.note||"")+(count?` · Aktuell ${count} Angebote mit konkret beziffertem Vorteil.`:" · Im aktuellen Abruf ist kein konkreter Euro-Preisvorteil ausgewiesen."); label.append(input,text); list.appendChild(label);} $("loyaltyAll").onclick=()=>{for(const input of list.querySelectorAll("input[data-program]")){input.checked=true;selectedPrograms.add(input.dataset.program)}syncLoyaltyUrl();query(true)}; $("loyaltyNone").onclick=()=>{for(const input of list.querySelectorAll("input[data-program]"))input.checked=false;selectedPrograms.clear();syncLoyaltyUrl();query(true)};}
+$("q").addEventListener("input",()=>{clearTimeout(debounce);debounce=setTimeout(()=>query(true),300)}); $("sort").addEventListener("change",()=>query(true)); document.querySelectorAll(".viewTab").forEach(button=>button.addEventListener("click",()=>{view=button.dataset.view||"best_only";query(true)}));
+new IntersectionObserver(e=>{if(e.some(x=>x.isIntersecting))query(false)},{rootMargin:"600px"}).observe($("sentinel")); query(true);
